Update provider design
